refactor(SearchModels): #402c-prime lift CandidateFetcher protocol + SmartCandidate + AvailabilityFilter
Slice C-prime of #402. Lifts three more types into SearchModels so the
last value-type/protocol-only consumer in Services (Sample.Services.CandidateFetcher)
can drop `import Search` entirely.

Types lifted from Search target to SearchModels:

1. Search.CandidateFetcher (protocol). Was in
   Packages/Sources/Search/CandidateFetcher.swift. Pure narrow protocol
   (`sourceName: String { get }` + one method); no behavioural body. The
   concrete fetchers (PackageFTSCandidateFetcher, AppleDocsCandidateFetcher,
   etc.) stay in the Search target and conform to the now-relocated
   protocol via their existing extensions — Search target imports
   SearchModels in those files.

2. Search.SmartCandidate (value type). Was in the same file. Pure
   data — source / identifier / title / chunk / rawScore / kind /
   metadata, all Sendable + Hashable.

3. Search.AvailabilityFilter (value type). Lifted out of
   Search.PackageQuery.AvailabilityFilter — was nested inside the
   PackageQuery actor, making it inaccessible to consumers that can't
   import the Search target. Now top-level under the Search namespace
   (`Search.AvailabilityFilter`); callsites across CLI / Search /
   Services updated. In CLI, the local CLI.Command.Search subcommand
   shadows `Search`, so the fully-qualified `SearchModels.Search.AvailabilityFilter`
   form is used at 7 sites in CLI.Command.PackageSearch.swift +
   CLI.Command.Search.SmartReport.swift.

Source target changes:

- Packages/Sources/Search/CandidateFetcher.swift: delete the `extension
  Search { ... SmartCandidate ... CandidateFetcher ... }` block (lines
  21-77). The PackageFTSCandidateFetcher + AppleDocsCandidateFetcher
  extensions below it stay; they reference the relocated types via the
  already-present `import SearchModels`.
- Packages/Sources/Search/Search.PackageQuery.swift: delete the nested
  `public struct AvailabilityFilter` declaration (lines 94-108).
  References inside the actor (`availability: AvailabilityFilter?`)
  still compile because the lifted Search.AvailabilityFilter is
  reachable via the imported SearchModels.

Services side:

- Packages/Sources/Services/Sample.Services.CandidateFetcher.swift
  drops `import Search` — its only Search.* references are now
  Search.CandidateFetcher (protocol it conforms to), Search.SmartCandidate
  (value type it constructs), and Search.AvailabilityFilter (filter
  it stores), all in SearchModels.

After this PR, Services' production `import Search` count drops from 8
files to 7 (just the actor-adjacent ReadCommands/* services +
ServiceContainer). The remaining 7 require the Search.Index protocol
seam (slice C2).

Verification: xcrun swift build clean; make test-clean -> 1414 tests
in 157 suites pass.

Part 3 of #402.
