+++ Packages/Sources/CLI/Commands/CLI.Command.PackageSearch.swift
@@ -78,10 +78,10 @@ extension CLI.Command {
             // to calling `PackageQuery.answer` directly — but it goes through the
             // exact same code path as `cupertino ask`, which means future ranking
             // tweaks land in one place.
-            let availabilityFilter: SearchModule.PackageQuery.AvailabilityFilter?
+            let availabilityFilter: SearchModels.Search.AvailabilityFilter?
             switch (platform, minVersion) {
             case let (platform?, minVersion?):
-                availabilityFilter = SearchModule.PackageQuery.AvailabilityFilter(
+                availabilityFilter = SearchModels.Search.AvailabilityFilter(
                     platform: platform,
                     minVersion: minVersion
                 )
