+++ Packages/Sources/Search/Search.PackageQuery.swift
@@ -91,22 +91,6 @@ extension Search {
             return String(cString: ptr)
         }
 
-        /// Optional platform-availability filter (#220).
-        /// `platform` is one of `iOS`, `macOS`, `tvOS`, `watchOS`, `visionOS`
-        /// (case-insensitive). `minVersion` is a dotted decimal like
-        /// `"16.0"` or `"10.15"`. Both must be set to filter; otherwise the
-        /// flag is ignored. NULL `min_<platform>` rows in `package_metadata`
-        /// are dropped when a filter is active (no annotation = unknown =
-        /// excluded from a platform-specific query).
-        public struct AvailabilityFilter: Sendable {
-            public let platform: String
-            public let minVersion: String
-            public init(platform: String, minVersion: String) {
-                self.platform = platform
-                self.minVersion = minVersion
-            }
-        }
-
         public func answer(
             _ question: String,
             maxResults: Int = 3,
